Fixed a few bugs related the naming convention.
I have changed the naming convention in the YAML
files, for tuples, files, etc, for the sake of the
potential GUI development to enritch the table
of properties with useful types.
This has introduced bugs in many places..., need to
re-test and keep in mind.
Started adding a XFORM property for all Elements, to
allow to re-position an element.
TODO:
- check if file exists (exp_name_xform)
- if file exists, a dictionary with the last xform will
  be loaded into self.config.xform, and can be queried maybe
  when entering the state (at the element level) and try
  to use it there to set a 2d or 3d transform.
- Regardless of the file, if the condition in the config. file says
  that the xform can be edited, I will implement a basic 2D mouse
  drag/scale/rotate that will be saved using a keystroke (s) and
  will be automatically set next time is loaded.